fix: setup of DishWasher

+++ app/src/main/java/com/gemini/energy/wordDocGenerator/SorterForWordDocumentGenerator.kt
@@ -674,7 +674,7 @@ class SorterForWordDocumentGenerator {
                             "Dishwasher",
                             single.energyPowerChange(),
                             costElectricity,
-                            single.age,
+                            single.age.toDouble(),
                             materialCost,
                             implementationCost,
                             savings,
